chore: 调整group目录
