--- src/components/_internal/group/src/index.vue
@@ -1,40 +1,0 @@
-<template>
-  <div class="datav-wrapper">
-    Group
-  </div>
-</template>
-
-<script lang='ts'>
-import { defineComponent, PropType, computed, toRef } from 'vue'
-import { useDataCenter, getFieldMap } from '@/components/_mixins/use-data-center'
-import { useApiStore } from '@/store/api'
-import { Group } from './group'
-
-export default defineComponent({
-  name: 'VGroup',
-  props: {
-    com: {
-      type: Object as PropType<Group>,
-      required: true,
-    },
-  },
-  setup(props) {
-    const apiStore = useApiStore()
-    useDataCenter(props.com)
-
-    const dv_data = computed(() => {
-      return apiStore.dataMap[props.com.id]?.source ?? {}
-    })
-
-    const dv_field = computed(() => {
-      return getFieldMap(props.com.apis.source.fields)
-    })
-
-    const config = toRef(props.com, 'config')
-    const attr = toRef(props.com, 'attr')
-
-    return {
-    }
-  },
-})
-</script>
